Variance Soft Shadow Mapping has been added
VSSM has been added in our framework. Because SAT suffers from precision
issues, we have added an option to visualize the expected visual quality
(in this case, without SAT), and another option to evaluate the
performance achieved with the SAT implementation. See keyboard() and
SoftShadow.frag  for details.

Moreover, we take advantage from the Multiple Render Target (MRT) to
render the G-Buffers used for Monte-Carlo sampling.